New Authentication added

+++ src/main/java/com/cg/census/profile/CensusProfiling/controller/AdminController.java
@@ -14,8 +14,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.cg.census.profile.CensusProfiling.exception.NoAccessException;
 import com.cg.census.profile.CensusProfiling.model.User;
 import com.cg.census.profile.CensusProfiling.model.UserFamilyMember;
+import com.cg.census.profile.CensusProfiling.secure.service.AppUserService;
 import com.cg.census.profile.CensusProfiling.service.AdminService;
 
 import io.swagger.annotations.ApiOperation;
@@ -31,99 +33,143 @@ public static final Logger LOG = LoggerFactory.getLogger(User.class);
 	@Autowired
 	private AdminService service ;
 	
-	///Admin can add Users
-	@ApiOperation(value = "Add New user" , authorizations = { @Authorization(value = "jwtToken")})
+	@Autowired
+	AppUserService appUserService;
+	
+	/// Admin can add Users
 	@PostMapping("/addUser")
 	public User userRegister(@RequestBody User user) {
 		LOG.info("addUser");
+		if (appUserService.loginStatus().getRole().toString().equals("ADMIN")) {
 		return service.userRegister(user);
+		}
+		else {
+			throw new NoAccessException("You dont have access");
+		}
 	}
 	
 	////Get All Users
-	@ApiOperation(value = "Get All Users" , authorizations = { @Authorization(value = "jwtToken")})
 	@GetMapping("/getAllUsers")
 	public List<User> getAllUsers() {
 		LOG.info("AllUsers");
+		if (appUserService.loginStatus().getRole().toString().equals("ADMIN")) {
 		return service.findAllUsers();
+		}
+		else {
+			throw new NoAccessException("You dont have access");
+		}
 	}
 	
-	@ApiOperation(value = "Get All Members added by Users" , authorizations = { @Authorization(value = "jwtToken")})
 	@GetMapping("/getAllFamilyMembers")
 	public List<UserFamilyMember> getAllFamilyMembers() {
 		LOG.info("AllUsers");
+		if (appUserService.loginStatus().getRole().toString().equals("ADMIN")) {
 		return service.getAllFamilyMembers();
+		}
+		return null;
 	}
 	
 	
 	///////Get Users by First Name
-	@ApiOperation(value = "Search Users by First Name" , authorizations = { @Authorization(value = "jwtToken")})
-	@GetMapping("/getbyfirstname/{firstName}")
+	@GetMapping("/getUserByfirstname/{firstName}")
 	public List<User> getByFirstName(@PathVariable("firstName") String firstName) {
 		LOG.info("getByFirstName");
+		if (appUserService.loginStatus().getRole().toString().equals("ADMIN")) {
 		return service.findUserByFirstName(firstName);
+		}
+		else {
+			throw new NoAccessException("You dont have access");
+		}
 	}
 	
 	/////Get Users by User ID
-	@ApiOperation(value = "Search User by ID" , authorizations = { @Authorization(value = "jwtToken")})
 	@GetMapping("/getUserById/{id}")
 	public User getUserById(@PathVariable("id") int id) {
 		LOG.info("getByFirstName");
+		if (appUserService.loginStatus().getRole().toString().equals("ADMIN")) {
 		return service.getUserById(id);
+		}
+		else {
+			throw new NoAccessException("You dont have access");
+		}
 	}
 	////Get Users by Last Name
-	@ApiOperation(value = "Search Users by using Lat Name filter" , authorizations = { @Authorization(value = "jwtToken")})
-	@GetMapping("/getUserByLastName/{lastName}")
+	@GetMapping("/getUsersByLastName/{lastName}")
 	public List<User> getByLastName(@PathVariable("lastName") String lastName) {
 		LOG.info("getBylastName");
+		if (appUserService.loginStatus().getRole().toString().equals("ADMIN")) {
 		return service.findUserByFirstName(lastName);
+		}
+		else {
+			throw new NoAccessException("You dont have access");
+		}
 	}
 
-	@ApiOperation(value = "Get all Users of given Age" , authorizations = { @Authorization(value = "jwtToken")})
 	@GetMapping("/getUsersByAge/{age}")
 	public List<User> getUsersByAge(@PathVariable("age") int age) {
 		LOG.info("getByFirstName");
+		if (appUserService.loginStatus().getRole().toString().equals("ADMIN")) {
 		return service.getUsersByAge(age);
+		}
+		else {
+			throw new NoAccessException("You dont have access");
+		}
 	}
 	
 	//Get user by Email
-	@ApiOperation(value = "Search Users by Email filter" , authorizations = { @Authorization(value = "jwtToken")})
 	@GetMapping("/getUserByEmail/{email}")
 	public List<User> getByEmail(@PathVariable("email") String email) {
 		LOG.info("getByEmail");
+		if (appUserService.loginStatus().getRole().toString().equals("ADMIN")) {
 		return service.findUserByEmail(email);
+		}
+		else {
+			throw new NoAccessException("You dont have access");
+		}
 	}
 	
-	@ApiOperation(value = "Search Users by City filter" , authorizations = { @Authorization(value = "jwtToken")})
 	@GetMapping("/getUsersByCity/{City}")
 	public List<User> getUserByCity(@PathVariable("City") String city) {
 		LOG.info("getByEmail");
+		if (appUserService.loginStatus().getRole().toString().equals("ADMIN")) {
 		return service.getUsersByCity(city);
+		}
+		else {
+			throw new NoAccessException("You dont have access");
+		}
 	}
 	
 	
 	///Admin can get All users by their Gender
-	@ApiOperation(value = "Search Users by Gender filter" , authorizations = { @Authorization(value = "jwtToken")})
 	@GetMapping("/getbyeGender/{gender}")
 	public List<User> getByGender(@PathVariable("gender") String gender) {
 		LOG.info("getByEmail");
+		if (appUserService.loginStatus().getRole().toString().equals("ADMIN")) {
 		return service.findUserByGender(gender);
-
+		}
+		else {
+			throw new NoAccessException("You dont have access");
+		}
 	}
 	
 	////Admin Can Delete User by ID
 	
 	@Transactional
-	@ApiOperation(value = "Delete User by ID" , authorizations = { @Authorization(value = "jwtToken")})
 	@DeleteMapping("/deleteUserById/{id}")
-	public void deleteUserById(@PathVariable("id") int id) {
-		service.deleteUserById(id);
+	public void deleteUsersById(@PathVariable("id") int id) {
+		LOG.info("Delte User by ID");
+		if (appUserService.loginStatus().getRole().toString().equals("ADMIN")) {
+			service.deleteUserById(id);
+		}
 	}
 	
 	@Transactional
-	@ApiOperation(value = "Delete User by First Name" , authorizations = { @Authorization(value = "jwtToken")})
 	@DeleteMapping("/deleteUserByFirstName/{name}")
 	public void deleteUserById(@PathVariable("name") String name) {
-		service.deleteUserByfirstName(name);
+		LOG.info("Delte User by First Name");
+		if (appUserService.loginStatus().getRole().toString().equals("ADMIN")) {
+			service.deleteUserByfirstName(name);
+		}
 	}
 	
 //	@ApiOperation(value = "Update Users information" , authorizations = { @Authorization(value = "jwtToken")})
