+++ src/main/java/com/cg/census/profile/CensusProfiling/controller/UserController.java
@@ -13,12 +13,11 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.cg.census.profile.CensusProfiling.exception.NoAccessException;
 import com.cg.census.profile.CensusProfiling.model.User;
 import com.cg.census.profile.CensusProfiling.model.UserFamilyMember;
+import com.cg.census.profile.CensusProfiling.secure.service.AppUserService;
 import com.cg.census.profile.CensusProfiling.service.UserService;
-
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.Authorization;
 
 @RestController
 public class UserController {
@@ -28,48 +27,67 @@ public class UserController {
 	@Autowired
 	private UserService service ;
 	
+	
+	@Autowired
+	AppUserService appUserService;
+	
 	////User Registration
-	//@ApiOperation(value = "User registration" , authorizations = { @Authorization(value = "jwtToken")})
-	@PostMapping("/register")
+	
+	@PostMapping("/registerUser")
 	public User userRegister(@RequestBody User user) {
-		LOG.info("addUser");
-		return service.userRegister(user);
+		LOG.info("User Register");
+			return service.userRegister(user);
 	}
 	
 //	@Autowired
 //	private UserFamilyMemberService service;
 
 	////Add New Family Member
-	@ApiOperation(value = "Add new Family Member" , authorizations = { @Authorization(value = "jwtToken")})
 	@PostMapping("/addmember")
 	public UserFamilyMember regMember(@RequestBody UserFamilyMember user) {
 		LOG.info("Member add");
-		return service.addMember(user);
+		if (appUserService.loginStatus().getRole().toString().equals("USER")) {
+			return service.addMember(user);
+		}
+		else {
+			throw new NoAccessException("You dont have access");
+		}
 	}
 	
 	////Delete Family Member by Name
-	@ApiOperation(value = "Delete Family Member using First Name" , authorizations = { @Authorization(value = "jwtToken")})
 	@DeleteMapping("/deleteMemberByFirstName/{name}")
 	public void deleteMember(@PathVariable(value = "name") String name) {
-		service.deleteMember(name);
+		LOG.info("Delete Family Member by First name");
+		if (appUserService.loginStatus().getRole().toString().equals("USER")) {
+			service.deleteMember(name);
+		}else {
+			throw new NoAccessException("You dont have access");
+		}
 	}
 
 	////Delete Family Member BY ID
-	@ApiOperation(value = "Delete Family Member using Member_ID" , authorizations = { @Authorization(value = "jwtToken")})
 	@DeleteMapping("/deleteMemberById/{mem_id}")
 	public void deleteMember(@PathVariable("mem_id") int mem_id) {
-		LOG.info("DeleteMember");
-		service.deleteMemberById(mem_id);
-		// This is delete
+		LOG.info("Delete Member Using ID");
+		if (appUserService.loginStatus().getRole().toString().equals("USER")) {
+			service.deleteMemberById(mem_id);
+		}else {
+			throw new NoAccessException("You dont have access");
+		}
+		
 	}
 	
 	
 	////Get Member by First Name
-	@ApiOperation(value = "Search Family Member using First Name filter" , authorizations = { @Authorization(value = "jwtToken")})
 	@GetMapping("/getMemberByFirstName/{fname}")
 	public List<UserFamilyMember> getMemberBYFirstName(@PathVariable(value = "fname") String fname) {
-		LOG.info("Get Member by FIrst Name");
-		return service.findMemberByFirstName(fname);
+		LOG.info("Get Member by First Name");
+		if (appUserService.loginStatus().getRole().toString().equals("USER")) {
+			return service.findMemberByFirstName(fname);
+		}
+		else {
+			throw new NoAccessException("You dont have access");
+		}
 	}
 	
 //	////Get Member by Last Name
@@ -80,43 +98,63 @@ public class UserController {
 //	}
 	
 	///Get Member by Id
-	@ApiOperation(value = "Search Family Member using Id filter" , authorizations = { @Authorization(value = "jwtToken")})
 	@GetMapping("/getMemberById/{mid}")
 	public List<UserFamilyMember> getMemberById(@PathVariable(value = "mid") int mid) {
 		LOG.info("Getting Family Member by ID");
+		if (appUserService.loginStatus().getRole().toString().equals("USER")) {
 		return service.findMemberById(mid);
+		}
+		else {
+			throw new NoAccessException("You dont have access");
+		}
 	}
 
 	////Get Member By Relation
-	@ApiOperation(value = "Search Family Member Using relation filter" , authorizations = { @Authorization(value = "jwtToken")})
 	@GetMapping("/getMemberByRelation/{relation}")
 	public List<UserFamilyMember> getMemberByRelation(@PathVariable(value = "relation") String relation) {
 		LOG.info("Getting Family Members By Relation");
+		if (appUserService.loginStatus().getRole().toString().equals("USER")) {
 		return service.findByRelation(relation);
+		}
+		else {
+			throw new NoAccessException("You dont have access");
+		}
 	}
 	
 	///Get Family Members using Date of birth
-	@ApiOperation(value = "Search Family Memeber using Age filter" , authorizations = { @Authorization(value = "jwtToken")})
-	@GetMapping("/getMembersByAge/{age}")
+	@GetMapping("/getAllMembersByAge/{age}")
 	public List<UserFamilyMember> getMemberByAge(@PathVariable(value = "age") int age) {
 		LOG.info("Getting Family Members By Relation");
-		return service.findByAge(age);
+		if (appUserService.loginStatus().getRole().toString().equals("ADMIN")) {
+			return service.findByAge(age);
+		}
+		else {
+			throw new NoAccessException("You dont have access");
+		}
 	}
 	
 	
-	@ApiOperation(value = "Update Users information" , authorizations = { @Authorization(value = "jwtToken")})
 	@PutMapping("/updateUserInfo/{id}")
 	public User updateUserProfile(@PathVariable("id") int id , @RequestBody User user) {
 		LOG.info("Update User Info");
+		if (appUserService.loginStatus().getRole().toString().equals("USER")) { 
 		return service.updateUserProfile(id, user);
+		}
+		else {
+			throw new NoAccessException("You dont have access");
+		}
 	}
 	
 	//// Update User's Family Member
 	@PutMapping("/updateFamilyMemInfo/{id}")
-	@ApiOperation(value = "Update information of Family Member" , authorizations = { @Authorization(value = "jwtToken")})
 	public UserFamilyMember updateMemId(@PathVariable("id") int id , @RequestBody UserFamilyMember memId) {
 		LOG.info("update memId");
+		if (appUserService.loginStatus().getRole().toString().equals("USER")) {
 		return service.updateMemberInfo(id, memId);
+		}
+		else {
+			throw new NoAccessException("You dont have access");
+		}
 	}
 
 }
